+++ src/main/java/com/nephest/battlenet/sc2/model/local/dao/DivisionDAO.java
@@ -88,7 +88,7 @@ public class DivisionDAO
 
     public List<Division> findListByLadder
     (
-        Long season, Region region,
+        Integer season, Region region,
         BaseLeague.LeagueType leagueType, QueueType queueType, TeamType teamType,
         BaseLeagueTier.LeagueTierType tierType
     )
